refactor: 프로젝트 모달 컴포넌트 분리 리스트 아이템에 들어갈 버튼들을 아이템으로 분리
